requirements.txt
Flask==2.3.4
PyPDF2==3.1.1
google-generativeai==0.14.1
gunicorn
